move handleGetDataClick up to app so both tabs can use it
